+++ src/app/ui/split-ui/components.ts
@@ -1,6 +1,7 @@
 import { ControlsWrapper, Log } from 'molstar/lib/mol-plugin-ui/plugin';
 import { SequenceView } from 'molstar/lib/mol-plugin-ui/sequence';
 import { JSXElementConstructor } from 'react';
+import { FullLayoutNoControlsUnlessExpanded } from '../layout-no-controls-unless-expanded';
 import { DefaultLeftPanelControls, PDBeLeftPanelControls } from '../left-panel/pdbe-left-panel';
 import { PDBeViewport } from '../pdbe-viewport';
 
@@ -23,6 +24,9 @@ export const UIComponents = {
 
     /** Component containing right panel controls (contents depend on PDBeMolstar init params (superposition/ligand/default view)) */
     DefaultRightPanelControls: ControlsWrapper,
+
+    /** Component containing only `PDBeViewport` when not in fullscreen view, but shows full UI layout when in fullscreen view. */
+    FullLayoutNoControlsUnlessExpanded,
 
     // TODO add all meaningful components,
 } as const satisfies Record<string, JSXElementConstructor<any>>;